[ECP-8992] Refactor code duplication
